backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
pillow
numpy
opencv-python
torch
torchvision
git+https://github.com/facebookresearch/segment-anything-2.git
transformers
pymatting